+++ Arkanoid/Engine/SDL/SDLRenderer.h
@@ -27,8 +27,8 @@ public:
 	virtual void DrawRect(const FRect& Rect) override;
 	virtual void FillRect(const FRect& Rect) override;
 
-	virtual void DrawCircle(const FPoint& Center, int32 Radius) override;
-	virtual void FillCircle(const FPoint& Center, int32 Radius) override;
+	virtual void DrawCircle(const FPoint& Center, float Radius) override;
+	virtual void FillCircle(const FPoint& Center, float Radius) override;
 
 	virtual void* GetNativeRenderer() const override { return NativeRenderer; }
